requirements.txt
Flask>=2.0
gunicorn
google-generativeai
python-dotenv

